backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
google-generativeai==0.3.2
reportlab==4.0.8
requests==2.31.0
